Safari-609.1.11.1/LayoutTests/editing/caret/ios/absolute-caret-position-after-scroll-expected.txt

PASS initialCaretRect.left is 6
PASS initialCaretRect.top is 21
PASS initialCaretRect.width is 3
PASS initialCaretRect.height is 15
PASS finalCaretRect.left is 6
PASS finalCaretRect.top is 21
PASS finalCaretRect.width is 3
PASS finalCaretRect.height is 15
PASS successfullyParsed is true

TEST COMPLETE
